requirements.txt
flask
python-dotenv
python-binance
cryptography
